refactor: split the worktree.changes wire contract out of the payloads module
remote-orchestrator-payloads.ts crossed the file-size cap. The seam is real:
worktree.changes is the one channel whose payload says two different things
about a key (counts, and the paths it could not read), so its parser and map
equality now live together and the rest of that module stays one shape per
channel. The test file follows the same line — pure decoders out, orchestrator
plumbing stays.
