+++ packages/kobe/src/client/remote-orchestrator-payloads.ts
@@ -63,6 +63,16 @@ export type EngineTabStateMap = ReadonlyMap<string, ReadonlyMap<string, TaskEngi
 /** Durable daemon-owned attention episode, pushed as a full snapshot. */
 export type AttentionInboxItem = ChannelPayloads["attention.inbox"]["items"][number]
 
+// The `worktree.changes` wire contract lives in its own module (its payload
+// carries two facts per key); re-exported here so existing importers keep
+// naming it through this one.
+import type { WorktreeChangesMap } from "./remote-orchestrator-worktree-changes.ts"
+export {
+  type WorktreeChangesMap,
+  parseWorktreeChangesPayload,
+  sameWorktreeChangesMap,
+} from "./remote-orchestrator-worktree-changes.ts"
+
 /**
  * A long daemon operation currently IN FLIGHT for a task, accumulated from
  * the `task.jobs` channel (today: `ensureWorktree` — `git worktree add` is
@@ -74,20 +84,6 @@ export type AttentionInboxItem = ChannelPayloads["attention.inbox"]["items"][num
 export interface TaskJobState {
   readonly kind: "ensureWorktree"
 }
-
-/**
- * Daemon-collected `+N −M` counts keyed by worktree path, from the
- * `worktree.changes` channel (one collector in the daemon
- * instead of per-pane git polling). `null` means "no daemon-collected
- * data": either the daemon predates the channel (absent from
- * `hello.capabilities`) or `init()` hasn't completed — the sidebar then
- * falls back to its local poller.
- */
-/** Path → counts, or `null` for a worktree the daemon TRACKED but could not
- *  read. A `null` entry and an absent key are different facts: absent means
- *  "not collected" (remote project, just-created task) and draws no chip;
- *  `null` means the read failed and draws the unknown mark. */
-export type WorktreeChangesMap = ReadonlyMap<string, WorktreeChanges | null>
 
 /**
  * Compact, bounded description of a dropped event payload for `client.log` —
@@ -107,37 +103,6 @@ export function describePayload(value: unknown): string {
   }
   if (text.length > 120) text = `${text.slice(0, 120)}…`
   return `${type}:${text}`
-}
-
-/**
- * Parse a `worktree.changes` wire payload into a path→counts map.
- * Returns `null` for a malformed payload (the event is then ignored —
- * never clobber a good map with garbage). Exported for unit tests.
- */
-export function parseWorktreeChangesPayload(payload: unknown): Map<string, WorktreeChanges | null> | null {
-  const changes = (payload as { changes?: unknown } | undefined)?.changes
-  if (!changes || typeof changes !== "object" || Array.isArray(changes)) return null
-  const map = new Map<string, WorktreeChanges | null>()
-  for (const [path, value] of Object.entries(changes as Record<string, unknown>)) {
-    const counts = value as { added?: unknown; deleted?: unknown; behind?: unknown; ahead?: unknown } | undefined
-    if (typeof counts?.added !== "number" || typeof counts.deleted !== "number") return null
-    // `behind` and `ahead` are both additive: an older daemon omits them, and
-    // the chip then simply does not draw — never a fabricated zero.
-    map.set(path, {
-      added: counts.added,
-      deleted: counts.deleted,
-      ...(typeof counts.behind === "number" ? { behind: counts.behind } : {}),
-      ...(typeof counts.ahead === "number" ? { ahead: counts.ahead } : {}),
-    })
-  }
-  // Worktrees the daemon tracked but could not read. Additive, so an older
-  // daemon that omits the field simply contributes no unknowns. Mapped to
-  // `null` — NOT left absent, which the row would draw as a clean chip.
-  const unreadable = (payload as { unreadable?: unknown } | undefined)?.unreadable
-  if (Array.isArray(unreadable)) {
-    for (const path of unreadable) if (typeof path === "string" && path.length > 0) map.set(path, null)
-  }
-  return map
 }
 
 /**
@@ -167,20 +132,6 @@ export function decodeUiPrefsPayload(payload: unknown): UiPrefsPayload | null {
     keysCollapsed: p.keysCollapsed === true,
     projectFilter: typeof p.projectFilter === "string" && p.projectFilter.length > 0 ? p.projectFilter : null,
   }
-}
-
-/**
- * Entry-wise equality for two changes maps — an unchanged republish (e.g.
- * the bus replaying its last value across a reconnect) must not churn the
- * signal and re-render every sidebar row. Exported for unit tests.
- */
-export function sameWorktreeChangesMap(a: WorktreeChangesMap, b: WorktreeChangesMap): boolean {
-  if (a.size !== b.size) return false
-  for (const [path, counts] of a) {
-    if (!b.has(path)) return false
-    if (!sameWorktreeChanges(counts, b.get(path) ?? null)) return false
-  }
-  return true
 }
 
 /**
